More fattening

+++ src/EmbassyAirlines.Application/DTOs/AircraftDto.cs
@@ -2,7 +2,7 @@ namespace EmbassyAirlines.Application.Dtos;
 
 public sealed record AircraftDto(Guid Id, DateTime CreatedAt,
     string Registration, string Model, string Type, int EconomySeats,
-    int BusinessSeats, int FlightHours, int BasicEmptyWeight,
+    int BusinessSeats, float FlightHours, int BasicEmptyWeight,
     int MaximumZeroFuelWeight, int MaximumTakeoffWeight,
     int MaximumLandingWeight, int MaximumCargoWeight, int FuelOnboard,
     int FuelCapacity, int MinimumCabinCrew);
