+++ src/EmbassyAirlines.Application/DTOs/NewAircraftDto.cs
@@ -1,16 +1,17 @@
 using System.Diagnostics.CodeAnalysis;
 using EmbassyAirlines.Application.Errors;
 using EmbassyAirlines.Application.Interfaces;
+using EmbassyAirlines.Application.Validators;
 using Mediator;
 
 namespace EmbassyAirlines.Application.Dtos;
 
 public sealed record NewAircraftDto(string Registration, string Model,
-    string Type, int EconomySeats, int BusinessSeats, int FlightHours,
+    string Type, int EconomySeats, int BusinessSeats, float FlightHours,
     int BasicEmptyWeight, int MaximumZeroFuelWeight, int MaximumTakeoffWeight,
     int MaximumLandingWeight, int MaximumCargoWeight, int FuelOnboard,
     int FuelCapacity, int MinimumCabinCrew) : ICommand<AircraftDto>, IValidate
-    {
+{
     public bool IsValid([NotNullWhen(false)] out ValidationError? error)
     {
         var validator = new NewAircraftDtoValidator();
